refactor: split application helper types
