+++ src/CADShared/CADShared.projitems
@@ -142,6 +142,10 @@
       <FsFoxModule>Cad.Application</FsFoxModule>
       <FsFoxOrder>0280</FsFoxOrder>
     </Compile>
+    <Compile Include="$(MSBuildThisFileDirectory)Cad\Application\Documents\DocumentLockManagerExtension.cs">
+      <FsFoxModule>Cad.Application</FsFoxModule>
+      <FsFoxOrder>0281</FsFoxOrder>
+    </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Cad\Database\Files\DwgMark.cs">
       <FsFoxModule>Cad.Database</FsFoxModule>
       <FsFoxOrder>0290</FsFoxOrder>
@@ -396,6 +400,10 @@
     <Compile Include="$(MSBuildThisFileDirectory)Cad\Application\SystemVariables\DBmod.cs">
       <FsFoxModule>Cad.Application</FsFoxModule>
       <FsFoxOrder>0760</FsFoxOrder>
+    </Compile>
+    <Compile Include="$(MSBuildThisFileDirectory)Cad\Application\SystemVariables\DBmodEx.cs">
+      <FsFoxModule>Cad.Application</FsFoxModule>
+      <FsFoxOrder>0761</FsFoxOrder>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Cad\Editor\Commands\PostCmd.cs">
       <FsFoxModule>Cad.Editor</FsFoxModule>
